requirements.txt
Flask
Flask-Uploads
opencv-python
tensorflow
cassandra-driver
